replacing uint with size_t to fix compilation on windows

+++ SofaKernel/modules/SofaCore/SofaCore_test/objectmodel/BaseLink_test.cpp
@@ -61,7 +61,7 @@ TEST_F(SingleLink_test, checkAccess  )
 {
     ASSERT_EQ(m_link.get(), m_dst.get()) ;
     ASSERT_FALSE(m_link.empty()) ;
-    ASSERT_EQ(m_link.size(), uint(1)) ;
+    ASSERT_EQ(m_link.size(), size_t(1)) ;
 }
 
 TEST_F(SingleLink_test, checkIsSetPersistent  )
@@ -86,18 +86,18 @@ TEST_F(SingleLink_test, checkCounterLogic )
 TEST_F(SingleLink_test, checkMultiLink )
 {
     SingleLink<BaseObject, BaseObject, BaseLink::FLAG_MULTILINK > mlink ;
-    ASSERT_EQ(mlink.size(), uint(0)) ;
+    ASSERT_EQ(mlink.size(), size_t(0)) ;
     mlink.add(m_dst.get()) ;
-    ASSERT_EQ(mlink.size(), uint(1)) ;
+    ASSERT_EQ(mlink.size(), size_t(1)) ;
     mlink.add(m_dst.get()) ;
-    ASSERT_EQ(mlink.size(), uint(1)) ;
+    ASSERT_EQ(mlink.size(), size_t(1)) ;
 
     SingleLink<BaseObject, BaseObject, BaseLink::FLAG_NONE > slink ;
-    ASSERT_EQ(slink.size(), uint(0)) ;
+    ASSERT_EQ(slink.size(), size_t(0)) ;
     slink.add(m_dst.get()) ;
-    ASSERT_EQ(slink.size(), uint(1)) ;
+    ASSERT_EQ(slink.size(), size_t(1)) ;
     slink.add(m_dst.get()) ;
-    ASSERT_EQ(slink.size(), uint(1)) ;
+    ASSERT_EQ(slink.size(), size_t(1)) ;
 }
 
 TEST_F(SingleLink_test, getOwnerBase)
